+++ HISTORY.md
@@ -8,6 +8,7 @@ https://github.com/josdejong/mathjs
   `bignumber.js`.
 - Implemented support for polar coordinates to the `Complex` datatype.
   Thanks Finn Pauls (@finnp).
+- Implemented BigNumber support for functions `exp`, `log`, and `log10`.
 - Implemented conditional operator `a ? b : c` in expression parser.
 - Improved function `ifElse`, is now specified for special data types too.
 - Implemented `BigNumber` support for function `det`.